backend/requirements.txt
# Core FastAPI dependencies - minimal set for Railway deployment
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and Auth
supabase==2.3.0

# Security and JWT  
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client
httpx==0.25.2

# Rate Limiting and Monitoring
slowapi==0.1.9
prometheus-client==0.19.0

# Redis for caching
redis>=5.0.0

# RAG and AI dependencies
chromadb==0.4.22
sentence-transformers==2.2.2
openai==1.12.0
langchain==0.1.6
langchain-community==0.0.19
together==0.2.10

# Document processing
pypdf==4.0.1
python-docx==1.1.0
python-magic==0.4.27
aiofiles==23.2.0

# Text processing
tiktoken==0.5.2
nltk==3.8.1